# api/openapi/catalog.yaml
openapi: 3.1.0
info:
  title: AnimeEnigma Catalog API
  version: 1.0.0
  description: |
    Anime catalog API with on-demand data population from Shikimori.
    When searching, anime data is fetched from Shikimori and cached locally.
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /admin/anime:
    post:
      operationId: createAnime
      summary: Manually add anime (admin only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnimeRequest'
      responses:
        '201':
          description: Anime created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimeResponse'
        '403':
          $ref: './common.yaml#/components/responses/Forbidden'
...
components:
  schemas:
    Anime:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          description: Romaji/English title
        name_ru:
          type: string
          description: Russian title
        name_jp:
          type: string
          description: Japanese title
        description:
          type: string
        year:
          type: integer
        season:
          type: string
          enum: [winter, spring, summer, fall]
        status:
          type: string
          enum: [ongoing, released, announced]
        episodes_count:
          type: integer
        episode_duration:
          type: integer
          description: Average episode duration in minutes
        score:
          type: number
          format: float
        poster_url:
          type: string
          format: uri
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
        external_ids:
          $ref: '#/components/schemas/ExternalIDs'
        has_video:
          type: boolean
          description: Whether video sources are available
        video_sources:
          type: array
          items:
            $ref: '#/components/schemas/VideoSourceSummary'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    VideoSourceSummary:
      type: object
      properties:
        type:
          type: string
          enum: [minio, external]
        quality:
          type: string
          enum: [360p, 480p, 720p, 1080p]
        language:
          type: string
        subtitles:
          type: array
          items:
            type: string
    Genre:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        name_ru:
          type: string
    ExternalIDs:
      type: object
      properties:
        shikimori:
          type: string
        mal:
          type: string
        anilist:
          type: string
        anidb:
          type: string
    AnimeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Anime'
...
    CreateAnimeRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        name_ru:
          type: string
        name_jp:
          type: string
        description:
          type: string
        year:
          type: integer
        season:
          type: string
        status:
          type: string
        episodes_count:
          type: integer
        poster_url:
          type: string
        genre_ids:
          type: array
          items:
            type: string
        shikimori_id:
          type: string
          description: If provided, will fetch additional data from Shikimori
...
  securitySchemes:
    BearerAuth:
      $ref: './common.yaml#/components/securitySchemes/BearerAuth'